requirements.txt
ffmpeg
openai-whisper 
jiwer
certifi
setuptools-rust